drivers: Handle -Wvoid-pointer-to-int-cast warning

+++ drivers/soc/oplus/system/oplus_project/qcom/oplus_project.c
@@ -22,8 +22,8 @@
 
 #define SMEM_PROJECT    135
 
-#define UINT2Ptr(n)        (uint32_t *)(n)
-#define Ptr2UINT32(p)    (uint32_t)(p)
+#define UINT2Ptr(n)        (uintptr_t *)(n)
+#define Ptr2UINT32(p)    (uintptr_t)(p)
 
 #define PROJECT_VERSION            (0x1)
 #define PCB_VERSION                (0x2)
